Move workhorse out of __init__
